+++ src/time.ml
@@ -1086,7 +1086,7 @@ let with_tz tz t = Unary_op (With_tz tz, t)
 let pattern ?(years = []) ?(year_ranges = []) ?(months = [])
     ?(month_ranges = []) ?(days = []) ?(day_ranges = []) ?(weekdays = [])
     ?(weekday_ranges = []) ?(hours = []) ?(hour_ranges = []) ?(minutes = [])
-    ?(minute_ranges = []) ?(seconds = []) ?(second_ranges = []) () : t =
+    ?(minute_ranges = []) ?(seconds = []) ?(second_ranges = []) ?(ns_ranges = []) () : t =
   let years =
     try years @ Year_ranges.Flatten.flatten_list year_ranges
     with Range.Range_is_invalid -> invalid_arg "pattern: invalid year range"
@@ -1139,6 +1139,12 @@ let pattern ?(years = []) ?(year_ranges = []) ?(months = [])
       invalid_arg "pattern: not all minutes are valid"
     else if Stdlib.not (List.for_all (fun x -> 0 <= x && x < 60) seconds) then
       invalid_arg "pattern: not all seconds are valid"
+    else if Stdlib.not (List.for_all (fun x ->
+        match x with
+        | `Range_inc (x, y) -> x <= y
+        | `Range_exc (x, y) -> x < y
+      ) ns_ranges) then
+      invalid_arg "pattern: invalid ns ranges"
     else
       let years = Int_set.of_list years in
       let months = Int_set.of_list months in
@@ -1156,6 +1162,16 @@ let pattern ?(years = []) ?(year_ranges = []) ?(months = [])
           hours;
           minutes;
           seconds;
+          ns =
+            ns_ranges
+            |> List.map (fun x ->
+                match x with
+                | `Range_inc (x, y) -> Diet.Int.Interval.make x y
+                | `Range_exc (x, y) -> Diet.Int.Interval.make x (pred y)
+              )
+            |> List.fold_left (fun acc x ->
+                Diet.Int.add x acc
+              ) Diet.Int.empty
         }
 
 let month_day_ranges_are_valid_strict ~safe_month_day_range_inc day_ranges =