+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 
 - Swapped to using `inc_exc` for `hms_intervals` API
 
+- Added nanosecond support for `pattern` and `Points.t`
+
 ## Timedesc 0.3.1
 
 - Fixed `Timedesc.Span.make` handling of `ns` when `ns = Int.min_int`